refactor: moved BlogNavigationButtons to component
