+++ routes/[toys]/spinners.tsx
@@ -5,7 +5,7 @@ import { CustomHead } from "../../components/CustomHead.tsx";
 //? A HTML Link component to pre-format links and reduce boiletplate
 import { GradientLink } from "../../components/GradientLink.tsx";
 //? Navigation Buttons to go back to the previous page or to the next article
-import BlogNavigationButtons from "../../islands/BlogNavigationButtons.tsx";
+import BlogNavigationButtons from "../../components/BlogNavigationButtons.tsx";
 
 export default function Home() {
   return (
