+++ routes/blog.tsx
@@ -3,7 +3,7 @@ import { CustomHead } from "../components/CustomHead.tsx";
 //? Lateral text with theme switcher
 import { Base } from "../components/Base.tsx";
 //? Navigation Buttons to go back to the previous page or to the next page (optional)
-import BlogNavigationButtons from "../islands/BlogNavigationButtons.tsx";
+import BlogNavigationButtons from "../components/BlogNavigationButtons.tsx";
 //? Import the template for Blog Post summaries
 import { BlogPostSummary } from "../components/BlogPostSummary.tsx";
 //? All posts so far
