+++ routes/projects.tsx
@@ -4,7 +4,7 @@ import { Base } from "../components/Base.tsx";
 import { CustomHead } from "../components/CustomHead.tsx";
 import { GradientLink } from "../components/GradientLink.tsx";
 //? Navigation Buttons to go back to the previous page or to the next article
-import BlogNavigationButtons from "../islands/BlogNavigationButtons.tsx";
+import BlogNavigationButtons from "../components/BlogNavigationButtons.tsx";
 
 export default function Home() {
   return (
