+++ routes/toys.tsx
@@ -6,7 +6,7 @@ import { CarouselCard } from "../components/CarouselCard.tsx";
 //? Import CustomHead with appropriate metadata
 import { CustomHead } from "../components/CustomHead.tsx";
 //? Navigation Buttons to go back to the previous page or to the next page (optional)
-import BlogNavigationButtons from "../islands/BlogNavigationButtons.tsx";
+import BlogNavigationButtons from "../components/BlogNavigationButtons.tsx";
 
 export default function Home() {
   return (
