+++ routes/what-is-this.tsx
@@ -3,7 +3,7 @@ import { Base } from "../components/Base.tsx";
 //? Import CustomHead with appropriate metadata
 import { CustomHead } from "../components/CustomHead.tsx";
 //? Navigation Buttons to go back to the previous page or to the next page (optional)
-import BlogNavigationButtons from "../islands/BlogNavigationButtons.tsx";
+import BlogNavigationButtons from "../components/BlogNavigationButtons.tsx";
 
 export default function Home() {
   return (
